requirements.txt
discord.py
torch
torchvision
Pillow
scipy